app/requirements.txt
django>=4,<5
django-redis
gunicorn
httpx
psycopg[binary]
python-decouple
